commenting out test fusha brand

+++ screens/StorageScreen.js
@@ -1,5 +1,5 @@
 import * as FileSystem from 'expo-file-system'
-import React, { useEffect, useState ***REMOVED*** from 'react'
+import React, { useEffect, useState } from 'react'
 import {
   Alert,
   Dimensions,
@@ -7,18 +7,18 @@ import {
   SafeAreaView,
   StyleSheet,
   View
-***REMOVED*** from 'react-native'
-import { connect ***REMOVED*** from 'react-redux'
+} from 'react-native'
+import { connect } from 'react-redux'
 import LanguageStorageItem from '../components/list-items/LanguageStorageItem'
 import BackButton from '../components/standard/BackButton'
 import WahaButton from '../components/standard/WahaButton'
-import { removeDownload ***REMOVED*** from '../redux/actions/downloadActions'
+import { removeDownload } from '../redux/actions/downloadActions'
 import {
   activeDatabaseSelector,
   activeGroupSelector
-***REMOVED*** from '../redux/reducers/activeGroup'
-import { colors ***REMOVED*** from '../styles/colors'
-import { getLanguageFont ***REMOVED*** from '../styles/typography'
+} from '../redux/reducers/activeGroup'
+import { colors } from '../styles/colors'
+import { getLanguageFont } from '../styles/typography'
 
 function mapStateToProps (state) {
   return {
@@ -27,31 +27,31 @@ function mapStateToProps (state) {
     translations: activeDatabaseSelector(state).translations,
     font: getLanguageFont(activeGroupSelector(state).language),
     activeGroup: activeGroupSelector(state)
-  ***REMOVED***
-***REMOVED***
+  }
+}
 
 function mapDispatchToProps (dispatch) {
   return {
     removeDownload: lessonID => {
       dispatch(removeDownload(lessonID))
-    ***REMOVED***
-  ***REMOVED***
-***REMOVED***
+    }
+  }
+}
 
 function StorageScreen ({
   // Props passed from navigation.
-  navigation: { setOptions, goBack ***REMOVED***,
+  navigation: { setOptions, goBack },
   isRTL,
   database,
   translations,
   font,
   activeGroup,
   removeDownload
-***REMOVED***) {
+}) {
   //+ STATE
 
   // keeps track of storage size of each language's downloaded chapter 2s
-  const [storageObject, setStorageObject] = useState({***REMOVED***)
+  const [storageObject, setStorageObject] = useState({})
 
   // keeps track of total storage for all downloaded chapter 2s
   const [totalStorage, setTotalStorage] = useState(0)
@@ -61,27 +61,27 @@ function StorageScreen ({
   useEffect(() => {
     setOptions(getNavOptions())
     getAllStorageUsed()
-  ***REMOVED***, [])
+  }, [])
 
   //+ NAV OPTIONS
 
   function getNavOptions () {
     return {
       headerRight: isRTL
-        ? () => <BackButton onPress={() => goBack()***REMOVED*** />
+        ? () => <BackButton onPress={() => goBack()} />
         : () => <View></View>,
       headerLeft: isRTL
         ? () => <View></View>
-        : () => <BackButton onPress={() => goBack()***REMOVED*** />
-    ***REMOVED***
-  ***REMOVED***
+        : () => <BackButton onPress={() => goBack()} />
+    }
+  }
 
   //+ FUNCTIONS
 
   // gets the storage size in megabytes for all the downloaded chapter 2s for a specific language
   async function getStorageUsedByLanguage (language) {
     var storageUsed = 0
-    var regex = new RegExp('[a-z]{2***REMOVED***.[0-9]{1,2***REMOVED***.[0-9]{1,2***REMOVED***.[0-9]{1,2***REMOVED***.*')
+    var regex = new RegExp('[a-z]{2}.[0-9]{1,2}.[0-9]{1,2}.[0-9]{1,2}.*')
     return await FileSystem.readDirectoryAsync(FileSystem.documentDirectory)
       .then(async contents => {
         for (const item of contents) {
@@ -90,23 +90,23 @@ function StorageScreen ({
             if (item.slice(0, 2) === language) {
               await FileSystem.getInfoAsync(
                 FileSystem.documentDirectory + item,
-                {***REMOVED***
-              ).then(({ size ***REMOVED***) => {
+                {}
+              ).then(({ size }) => {
                 storageUsed += Math.round(size / 1000000)
-              ***REMOVED***)
-            ***REMOVED***
-          ***REMOVED***
-        ***REMOVED***
-      ***REMOVED***)
+              })
+            }
+          }
+        }
+      })
       .then(() => {
         return storageUsed
-      ***REMOVED***)
-  ***REMOVED***
+      })
+  }
 
   // get the storage size in megabytes of all downloaded chapter 2s
   function getAllStorageUsed () {
     setTotalStorage(0)
-    setStorageObject({***REMOVED***)
+    setStorageObject({})
     var languages = getInstalledLanguageInstances()
     for (const language of languages) {
       getStorageUsedByLanguage(language.languageID).then(storageUsed => {
@@ -114,57 +114,57 @@ function StorageScreen ({
         setStorageObject(oldObject => ({
           ...oldObject,
           [language.languageID]: storageUsed
-        ***REMOVED***))
-      ***REMOVED***)
-    ***REMOVED***
-  ***REMOVED***
+        }))
+      })
+    }
+  }
 
   // deletes all downloaded chapter 2s for a specific language
   // note: if no language specified, deletes all chapter 2s
   async function deleteDownloadedLessons (language) {
     await FileSystem.readDirectoryAsync(FileSystem.documentDirectory)
       .then(contents => {
-        var regex = new RegExp('[a-z]{2***REMOVED***.[0-9]{1,2***REMOVED***.[0-9]{1,2***REMOVED***.[0-9]{1,2***REMOVED***.*')
+        var regex = new RegExp('[a-z]{2}.[0-9]{1,2}.[0-9]{1,2}.[0-9]{1,2}.*')
         for (const item of contents) {
           var hasMatch = regex.exec(item)
           if (hasMatch) {
             if (!language) {
               FileSystem.deleteAsync(FileSystem.documentDirectory + item)
               removeDownload(item.slice(0, 5))
-            ***REMOVED*** else if (item.slice(0, 2) === language) {
+            } else if (item.slice(0, 2) === language) {
               FileSystem.deleteAsync(FileSystem.documentDirectory + item)
               removeDownload(item.slice(0, 5))
-            ***REMOVED***
-          ***REMOVED***
-        ***REMOVED***
-      ***REMOVED***)
+            }
+          }
+        }
+      })
       .then(() => {
         getAllStorageUsed()
-      ***REMOVED***)
-  ***REMOVED***
+      })
+  }
 
   // gets all the installed languages
   function getInstalledLanguageInstances () {
     var installedLanguageInstances = []
     for (key in database) {
       if (key.length === 2) {
-        var languageObject = {***REMOVED***
+        var languageObject = {}
         languageObject['languageName'] = database[key].displayName
         languageObject['languageID'] = key
         installedLanguageInstances.push(languageObject)
-      ***REMOVED***
-    ***REMOVED***
+      }
+    }
     return installedLanguageInstances
-  ***REMOVED***
+  }
 
   //+ RENDER
 
   function renderLanguageStorageItem (languageList) {
     return (
       <LanguageStorageItem
-        languageName={translations.general.brands[languageList.item.languageID]***REMOVED***
-        languageID={languageList.item.languageID***REMOVED***
-        megabytes={storageObject[languageList.item.languageID]***REMOVED***
+        languageName={translations.general.brands[languageList.item.languageID]}
+        languageID={languageList.item.languageID}
+        megabytes={storageObject[languageList.item.languageID]}
         clearDownloads={() => {
           Alert.alert(
             translations.storage.popups
@@ -174,36 +174,36 @@ function StorageScreen ({
             [
               {
                 text: translations.general.cancel,
-                onPress: () => {***REMOVED***
-              ***REMOVED***,
+                onPress: () => {}
+              },
               {
                 text: translations.general.ok,
                 onPress: () =>
                   deleteDownloadedLessons(languageList.item.languageID)
-              ***REMOVED***
+              }
             ]
           )
-        ***REMOVED******REMOVED***
+        }}
       />
     )
-  ***REMOVED***
+  }
 
   return (
-    <SafeAreaView style={styles.screen***REMOVED***>
+    <SafeAreaView style={styles.screen}>
       <FlatList
-        data={getInstalledLanguageInstances()***REMOVED***
-        renderItem={renderLanguageStorageItem***REMOVED***
-        keyExtractor={item => item.languageID***REMOVED***
+        data={getInstalledLanguageInstances()}
+        renderItem={renderLanguageStorageItem}
+        keyExtractor={item => item.languageID}
         ItemSeparatorComponent={() => (
-          <View style={{ height: 20, width: '100%' ***REMOVED******REMOVED*** />
-        )***REMOVED***
+          <View style={{ height: 20, width: '100%' }} />
+        )}
         ListHeaderComponent={() => (
-          <View style={{ height: 10, width: '100%' ***REMOVED******REMOVED*** />
-        )***REMOVED***
+          <View style={{ height: 10, width: '100%' }} />
+        )}
       />
       <WahaButton
         type='filled'
-        color={colors.red***REMOVED***
+        color={colors.red}
         label={
           translations.storage.clear_all_downloaded_lessons_button_label +
           ' (' +
@@ -211,8 +211,8 @@ function StorageScreen ({
           ' ' +
           translations.storage.megabyte_label +
           ')'
-        ***REMOVED***
-        width={Dimensions.get('window').width - 40***REMOVED***
+        }
+        width={Dimensions.get('window').width - 40}
         onPress={() =>
           Alert.alert(
             translations.storage.popups.clear_all_downloaded_lessons_title,
@@ -220,20 +220,20 @@ function StorageScreen ({
             [
               {
                 text: translations.general.cancel,
-                onPress: () => {***REMOVED***
-              ***REMOVED***,
+                onPress: () => {}
+              },
               {
                 text: translations.general.ok,
                 onPress: () => deleteDownloadedLessons()
-              ***REMOVED***
+              }
             ]
           )
-        ***REMOVED***
-        style={{ alignSelf: 'center' ***REMOVED******REMOVED***
+        }
+        style={{ alignSelf: 'center' }}
       />
     </SafeAreaView>
   )
-***REMOVED***
+}
 
 //+ STYLES
 
@@ -241,7 +241,7 @@ const styles = StyleSheet.create({
   screen: {
     flex: 1,
     backgroundColor: colors.aquaHaze
-  ***REMOVED***
-***REMOVED***)
+  }
+})
 
 export default connect(mapStateToProps, mapDispatchToProps)(StorageScreen)
